+++ tests/testthat/test-make_frames.r
@@ -1,42 +1,170 @@
-context("Check make_frames")
-
 # make example records
-pos1 <- structure(list(
-  animal_id = c("153", "153", "153", "153"),
-  bin_timestamp = structure(c(
-    1335704727,
-    1335704727, 1335791127, 1335877527
-  ), class = c("POSIXct", "POSIXt"), tzone = "UTC"), latitude = c(
-    43.60963, 43.61235, 43.6157011174478,
-    43.6404351467905
-  ), longitude = c(
-    -83.88658, -83.8608, -83.858826537583,
-    -83.8442607462938
-  ), record_type = c(
-    "detection", "detection",
-    "interpolated", "interpolated"
+pos <- interpolate_path(
+  walleye_detections[walleye_detections$animal_id == 153, ][
+    1:125,
+  ],
+  start_time = as.POSIXct("2012-04-29 13:05:27")
+)[1:4, ]
+
+
+test_that("Expected result when background lims and map not supplied", {
+  # Check output with default background_ylim, background_xlim, and bg_map
+  temp_dir <- tempdir()
+
+  expect_message(
+    make_frames(pos, out_dir = temp_dir, preview = TRUE),
+    "Preview frames written to"
   )
-), row.names = 4:7, class = "data.frame")
 
-# make preview image
-temp_dir <- tempdir()
-make_frames(pos1, out_dir = temp_dir, preview = TRUE)
+  # Check that file exists
+  expect_true(
+    file.exists(
+      file.path(temp_dir, "1.png")
+    )
+  )
 
-# Actual file sizes
-img_file <- file.path(temp_dir, "1.png")
-img_size <- file.info(img_file)$size
+  # Check file size greater than 30000 bytes
+  expect_gt(
+    file.size(
+      file.path(temp_dir, "1.png")
+    ),
+    30000
+  )
 
-# Expected file sizes
-size_should_be <- 30919
+  # Clean up
+  unlink(
+    list.files(
+      temp_dir,
+      full.names = TRUE,
+      recursive = TRUE,
+      include.dirs = TRUE
+    ),
+    recursive = TRUE
+  )
+})
 
-# Clean up
-unlink(list.files(temp_dir,
-  full.names = TRUE, recursive = TRUE,
-  include.dirs = TRUE
-), recursive = TRUE)
+test_that("Expected result when map but not background lims supplied", {
+  temp_dir <- tempdir()
 
-# Testing file size results
-test_that("making preview image expected result", {
-  # Check if expected and actual file sizes
-  expect_equal(img_size, size_should_be)
+  expect_message(
+    make_frames(
+      pos,
+      out_dir = temp_dir,
+      preview = TRUE,
+      bg_map = great_lakes_polygon
+    ),
+    "Preview frames written to"
+  )
+
+  # Check that file exists
+  expect_true(
+    file.exists(
+      file.path(temp_dir, "1.png")
+    )
+  )
+
+  # Check if file size greater than 20000 bytes
+  expect_gt(
+    file.size(
+      file.path(temp_dir, "1.png")
+    ),
+    20000
+  )
+
+  # Clean up
+  unlink(
+    list.files(
+      temp_dir,
+      full.names = TRUE,
+      recursive = TRUE,
+      include.dirs = TRUE
+    ),
+    recursive = TRUE
+  )
+})
+
+test_that("Expected result when map and background lims supplied", {
+  temp_dir <- tempdir()
+
+  expect_message(
+    make_frames(pos,
+      out_dir = temp_dir,
+      background_ylim = c(42, 47),
+      background_xlim = c(-90, -78),
+      preview = TRUE,
+      bg_map = great_lakes_polygon
+    ),
+    "Preview frames written to"
+  )
+
+  # Check that file exists
+  expect_true(
+    file.exists(
+      file.path(temp_dir, "1.png")
+    )
+  )
+
+  # Check if file size greater than 30000 bytes
+  expect_gt(
+    file.size(
+      file.path(temp_dir, "1.png")
+    ),
+    30000
+  )
+
+  # Clean up
+  unlink(
+    list.files(
+      temp_dir,
+      full.names = TRUE,
+      recursive = TRUE,
+      include.dirs = TRUE
+    ),
+    recursive = TRUE
+  )
+})
+
+test_that("Expected result when map is spatVector, bg lims not supplied", {
+  sv_poly <- terra::vect(great_lakes_polygon)
+
+  temp_dir <- tempdir()
+  expect_message(
+    expect_message(
+      make_frames(pos,
+        out_dir = temp_dir,
+        preview = TRUE,
+        bg_map = sv_poly
+      ),
+      "Converted terra object to sf"
+    ),
+    "Preview frames written to"
+  )
+
+  # Check that file exists
+  expect_true(
+    file.exists(
+      file.path(temp_dir, "1.png")
+    )
+  )
+
+
+  # Check if file size greater than 20000 bytes
+  expect_gt(
+    file.size(
+      file.path(temp_dir, "1.png")
+    ),
+    20000
+  )
+
+
+  # Clean up
+  unlink(
+    list.files(
+      temp_dir,
+      full.names = TRUE,
+      recursive = TRUE,
+      include.dirs = TRUE
+    ),
+    recursive = TRUE
+  )
 })